# packages/contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Pal Control API
  version: 0.4.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    Browser-facing API for Palworld operations. The official Palworld REST API
    and native mod bridge are internal adapters and must not be exposed directly.
    Management operations require a configured administrator API key and are
    additionally restricted to the loopback management boundary. Privileged
    economy and season writes also require a TOTP code and an auditable reason.
    Player-portal sessions use a separate cookie identity and never authorize
    management operations. Non-loopback deployments require HTTPS at a trusted
    reverse proxy in addition to these application controls.
servers:
  - url: http://127.0.0.1:5180/api/v1
    description: Local development
...
paths:
...
  /servers/{serverId}/players:
    get:
      tags: [Players]
      summary: List players
      operationId: listPlayers
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - name: q
          in: query
          schema:
            type: string
        - name: online
          in: query
          schema:
            type: boolean
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Player page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /servers/{serverId}/players/{playerId}:
    get:
      tags: [Players]
      summary: Get player details
      operationId: getPlayer
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PlayerId"
      responses:
        "200":
          description: Player details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /servers/{serverId}/paldefender/players:
    get:
      tags: [PalDefender]
      summary: List all PalDefender players
      operationId: listPalDefenderPlayers
      x-paldefender-permission: REST.Players.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
  /servers/{serverId}/paldefender/player/{playerIdentifier}:
    get:
      tags: [PalDefender]
      summary: Read one PalDefender player
      operationId: getPalDefenderPlayer
      x-paldefender-permission: REST.Player.Read
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/PalDefenderPlayerIdentifier"
      responses:
        "200":
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/PalDefenderUnavailable"
        default:
          $ref: "#/components/responses/PalDefenderUpstreamResponse"
...
  /extraction/admin/new-player-activities:
    get:
      tags: [ExtractionAdmin]
      summary: List immutable and draft new-player activity versions
      operationId: listNewPlayerActivityVersions
      parameters:
        - name: activityKey
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NewPlayerActivityKey"
      responses:
        "200":
          description: Activity versions visible to an EconomyAdmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewPlayerActivityList"
        default:
          $ref: "#/components/responses/ExtractionError"
  /extraction/admin/new-player-activities/{activityKey}/versions:
    post:
      tags: [ExtractionAdmin]
      summary: Create the next draft version of a new-player activity
      description: >-
        Always creates a draft with the next monotonically increasing version.
        Published and closed versions are never edited in place.
      operationId: createNewPlayerActivityDraft
      parameters:
        - $ref: "#/components/parameters/NewPlayerActivityKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPlayerActivityDefinition"
      responses:
        "201":
          description: Newly created draft version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewPlayerActivity"
        default:
          $ref: "#/components/responses/ExtractionError"
  /extraction/admin/new-player-activities/{activityKey}/versions/{version}:
    get:
      tags: [ExtractionAdmin]
      summary: Read one new-player activity version
      operationId: getNewPlayerActivityVersion
      parameters:
        - $ref: "#/components/parameters/NewPlayerActivityKey"
        - $ref: "#/components/parameters/NewPlayerActivityVersion"
      responses:
        "200":
          description: Requested version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewPlayerActivity"
        default:
          $ref: "#/components/responses/ExtractionError"
    put:
      tags: [ExtractionAdmin]
      summary: Update a draft new-player activity version
      description: Published and closed versions fail with a stable conflict code.
      operationId: updateNewPlayerActivityDraft
      parameters:
        - $ref: "#/components/parameters/NewPlayerActivityKey"
        - $ref: "#/components/parameters/NewPlayerActivityVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNewPlayerActivityDraftRequest"
      responses:
        "200":
          description: Updated draft version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewPlayerActivity"
        default:
          $ref: "#/components/responses/ExtractionError"
  /extraction/admin/new-player-activities/{activityKey}/versions/{version}/publish:
    post:
      tags: [ExtractionAdmin]
      summary: Publish a draft into the current weekly world
      description: >-
        High-risk EconomyAdmin operation. Publication freezes the active season
        id, exact world GUID, rewards, and content. A previous published version
        of the same activity key must be closed first.
      operationId: publishNewPlayerActivityVersion
      parameters:
        - $ref: "#/components/parameters/NewPlayerActivityKey"
        - $ref: "#/components/parameters/NewPlayerActivityVersion"
        - $ref: "#/components/parameters/AdminTotp"
        - $ref: "#/components/parameters/AdminReason"
      responses:
        "200":
          description: Published immutable version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewPlayerActivity"
        default:
          $ref: "#/components/responses/ExtractionError"
  /extraction/admin/new-player-activities/{activityKey}/versions/{version}/close:
    post:
      tags: [ExtractionAdmin]
      summary: Close a published new-player activity version
      description: High-risk EconomyAdmin operation recorded by the administrator audit middleware.
      operationId: closeNewPlayerActivityVersion
      parameters:
        - $ref: "#/components/parameters/NewPlayerActivityKey"
        - $ref: "#/components/parameters/NewPlayerActivityVersion"
        - $ref: "#/components/parameters/AdminTotp"
        - $ref: "#/components/parameters/AdminReason"
      responses:
        "200":
          description: Closed immutable version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewPlayerActivity"
        default:
          $ref: "#/components/responses/ExtractionError"
...
components:
  securitySchemes:
    adminApiKey:
      type: apiKey
      in: header
      name: X-Pal-Admin-Key
      description: >-
        Administrator API key. Configure only a SHA-256 digest at rest; never
        store the plaintext key in repository configuration or logs.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    playerSessionCookie:
      type: apiKey
      in: cookie
      name: pal_player_session
      description: >-
        Opaque player session set only after successful in-game OTP verification.
        The deployment uses an HttpOnly, SameSite=Strict cookie scoped to
        /api/v1/player and enables Secure behind HTTPS. Authenticated player
        identity is derived exclusively from this cookie, never from a userId
        query parameter or business request body.
  parameters:
    AdminTotp:
      name: X-Pal-Admin-Totp
      in: header
      required: true
      description: Six-digit RFC 6238 TOTP code required for privileged writes.
      schema:
        type: string
        pattern: "^[0-9]{6}$"
    AdminReason:
      name: X-Pal-Admin-Reason
      in: header
      required: true
      description: Human-entered audit reason for the privileged operation.
      schema:
        type: string
        minLength: 1
        maxLength: 512
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    PlayerId:
      name: playerId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    NewPlayerActivityKey:
      name: activityKey
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/NewPlayerActivityKey"
    NewPlayerActivityVersion:
      name: version
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 1000000
...
    PalDefenderPlayerIdentifier:
      name: playerIdentifier
      in: path
      required: true
      description: PalDefender accepts a supported player name, UserId, or platform identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 160
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unavailable:
      description: Required Palworld adapter is unavailable or in safe mode
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    PalDefenderUpstreamResponse:
      description: >-
        PalDefender upstream status, JSON, or text passed through synchronously.
        The Control API never exposes the upstream bearer token.
      content:
        application/json:
          schema: {}
        text/plain:
          schema:
            type: string
    PalDefenderUnavailable:
      description: PalDefender is disabled, timed out, or unavailable before a read completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PalDefenderTransportError"
...
    ExtractionError:
      description: Extraction-mode validation, state, persistence, or adapter error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    PalDefenderTransportError:
      type: object
      additionalProperties: false
      required: [error, outcomeUncertain]
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
        outcomeUncertain:
          type: boolean
          description: False for synchronous GET operations.
...
    PlayerPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlayerSummary"
        nextCursor:
          type: [string, "null"]
    PlayerSummary:
      type: object
      required: [playerId, name, online]
      properties:
        playerId:
          type: string
        uid:
          type: [string, "null"]
        name:
          type: string
        online:
          type: boolean
        level:
          type: [integer, "null"]
...
    Player:
      allOf:
        - $ref: "#/components/schemas/PlayerSummary"
        - type: object
          required: [sessionId]
          properties:
            sessionId:
              type: [string, "null"]
            guildId:
              type: [string, "null"]
            lastSeenAt:
              type: [string, "null"]
              format: date-time
...
    NewPlayerActivityKey:
      type: string
      minLength: 3
      maxLength: 64
      pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{2,63}$"
    NewPlayerActivityRewards:
      type: object
      additionalProperties: false
      required: [merchantCoin, weeklyTicket]
      properties:
        merchantCoin:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
        weeklyTicket:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
    NewPlayerActivityDefinition:
      type: object
      additionalProperties: false
      required: [title, description, marketCoin, seasonVoucher]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          minLength: 1
          maxLength: 1024
        marketCoin:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
        seasonVoucher:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
    UpdateNewPlayerActivityDraftRequest:
      type: object
      additionalProperties: false
      required: [title, description, marketCoin, seasonVoucher, expectedRevision]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          minLength: 1
          maxLength: 1024
        marketCoin:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
        seasonVoucher:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
        expectedRevision:
          type: integer
          format: int64
          minimum: 0
    NewPlayerActivity:
      type: object
      additionalProperties: false
      required:
        - activityId
        - activityKey
        - version
        - state
        - title
        - description
        - rewards
        - revision
        - createdBy
        - createdAt
        - publishedSeasonId
        - publishedWorldId
        - publishedBy
        - publishedAt
        - closedBy
        - closedAt
      properties:
        activityId:
          type: string
          format: uuid
        activityKey:
          $ref: "#/components/schemas/NewPlayerActivityKey"
        version:
          type: integer
          minimum: 1
        state:
          type: string
          enum: [draft, published, closed]
        title:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type: string
          minLength: 1
          maxLength: 1024
        rewards:
          $ref: "#/components/schemas/NewPlayerActivityRewards"
        revision:
          type: integer
          format: int64
          minimum: 0
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        publishedSeasonId:
          type: [string, "null"]
          format: uuid
        publishedWorldId:
          type: [string, "null"]
          pattern: "^[A-Fa-f0-9]{32}$"
        publishedBy:
          type: [string, "null"]
        publishedAt:
          type: [string, "null"]
          format: date-time
        closedBy:
          type: [string, "null"]
        closedAt:
          type: [string, "null"]
          format: date-time
    NewPlayerActivityList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NewPlayerActivity"
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: [string, "null"]
security:
  - adminApiKey: []